# models/openAPI.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AudiMeta API",
        "version": "1.0.2",
        "description": "API to retrieve information about Book, Series and Search"
    },
    "servers": [
        {
            "url": "https://audimeta.de/",
            "description": "Public Instance"
        }
    ],
    "paths": {
...
        "/series/books/{asin}": {
            "get": {
                "tags": [
                    "series"
                ],
                "summary": "Gets all books in a series",
                "parameters": [
                    {
                        "name": "region",
                        "in": "query",
                        "description": "The region to search in. Can be a comma separated list of max 3 regions. Defaults to US",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "US",
                                "UK",
                                "AU",
                                "DE",
                                "FR",
                                "CA",
                                "IT",
                                "JP",
                                "IN",
                                "BR",
                                "MX",
                                "ES"
                            ]
                        }
                    },
                    {
                        "name": "asin",
                        "in": "path",
                        "description": "The asin of the series",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books in series",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "asin": {
                                                "type": "string",
                                                "description": "The books asin"
                                            },
                                            "title": {
                                                "type": "string",
                                                "description": "The title of the book"
                                            },
                                            "regions": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "enum": [
                                                        "US",
                                                        "UK",
                                                        "AU",
                                                        "DE",
                                                        "FR",
                                                        "CA",
                                                        "IT",
                                                        "JP",
                                                        "IN",
                                                        "BR",
                                                        "MX",
                                                        "ES"
                                                    ]
                                                },
                                                "description": "The regions the book is available in"
                                            },
                                            "copyrigth": {
                                                "type": "string",
                                                "description": "The copyrigth of the book"
                                            },
                                            "description": {
                                                "type": "string",
                                                "description": "The description of the book"
                                            },
                                            "summary": {
                                                "type": "string",
                                                "description": "The summary of the book"
                                            },
                                            "explicit": {
                                                "type": "boolean",
                                                "description": "The explicit content of the book"
                                            },
                                            "isbn": {
                                                "type": "string",
                                                "description": "The isbn of the book"
                                            },
                                            "language": {
                                                "type": "string",
                                                "description": "The language of the book"
                                            },
                                            "publisher": {
                                                "type": "string",
                                                "description": "The publisher of the book"
                                            },
                                            "rating": {
                                                "type": "number",
                                                "description": "The rating of the book"
                                            },
                                            "series": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "asin": {
                                                            "type": "string",
                                                            "description": "The series asin"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "description": "The name of the series"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "description": "The description of the series"
                                                        },
                                                        "position": {
                                                            "type": "number",
                                                            "description": "The position of the series"
                                                        }
                                                    },
                                                    "required": ["asin", "name"]
                                                },
                                                "description": "The series of the book"
                                            },
                                            "authors": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "asin": {
                                                            "type": "string",
                                                            "description": "The author asin. Can be the authors name!"
                                                        },
                                                        "region": {
                                                            "type": "string",
                                                            "enum": [
                                                                "US",
                                                                "UK",
                                                                "AU",
                                                                "DE",
                                                                "FR",
                                                                "CA",
                                                                "IT",
                                                                "JP",
                                                                "IN",
                                                                "BR",
                                                                "MX",
                                                                "ES"
                                                            ],
                                                            "description": "The language of the authors description"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "description": "The name of the author"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "description": "The description of the author"
                                                        },
                                                        "image": {
                                                            "type": "string",
                                                            "description": "The image of the author. 512x512 resolution"
                                                        },
                                                        "genres": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "asin": {
                                                                        "type": "string",
                                                                        "description": "The genre asin"
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "description": "The name of the genre"
                                                                    },
                                                                    "type": {
                                                                        "type": "string",
                                                                        "enum": [
                                                                            "Genres",
                                                                            "Tags"
                                                                        ],
                                                                        "description": "The type of the genre"
                                                                    }
                                                                },
                                                                "required": ["asin", "name", "type"]
                                                            },
                                                            "description": "The genres of the author"
                                                        }
                                                    },
                                                    "required": ["asin", "region", "name"]
                                                },
                                                "description": "The authors of the book"
                                            },
                                            "narrators": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "description": "The name of the narrator"
                                                        }
                                                    },
                                                    "required": ["name"]
                                                },
                                                "description": "The narrators of the book"
                                            },
                                            "genres": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "asin": {
                                                            "type": "string",
                                                            "description": "The genre asin"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "description": "The name of the genre"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "enum": [
                                                                "Genres",
                                                                "Tags"
                                                            ],
                                                            "description": "The type of the genre"
                                                        }
                                                    },
                                                    "required": ["asin", "name", "type"]
                                                },
                                                "description": "The genres of the book"
                                            },
                                            "releaseDate": {
                                                "type": "string",
                                                "description": "The release date of the book"
                                            }
                                        },
                                        "required": ["asin", "title"]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/series/{asin}": {
            "get": {
                "tags": [
                    "series"
                ],
                "summary": "Gets information about a series",
                "parameters": [
                    {
                        "name": "region",
                        "in": "query",
                        "description": "The region to search in. Can be a comma separated list of max 3 regions. Defaults to US",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "US",
                                "UK",
                                "AU",
                                "DE",
                                "FR",
                                "CA",
                                "IT",
                                "JP",
                                "IN",
                                "BR",
                                "MX",
                                "ES"
                            ]
                        }
                    },
                    {
                        "name": "asin",
                        "in": "path",
                        "description": "The asin of the series",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Series information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "asin": {
                                            "type": "string",
                                            "description": "The series asin"
                                        },
                                        "name": {
                                            "type": "string",
                                            "description": "The name of the series"
                                        },
                                        "description": {
                                            "type": "string",
                                            "description": "The description of the series"
                                        },
                                        "position": {
                                            "type": "number",
                                            "description": "The position of the series"
                                        }
                                    },
                                    "required": ["asin", "name"]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Series not found"
                    }
                }
            }
        },
...
    }
}